cherry-pick 19398
