+++ src/frontend/src/expr/table_function.rs
@@ -17,7 +17,9 @@ use std::sync::{Arc, LazyLock};
 use itertools::Itertools;
 use risingwave_common::array::arrow::IcebergArrowConvert;
 use risingwave_common::types::{DataType, ScalarImpl, StructType};
-use risingwave_connector::source::iceberg::{create_parquet_stream_builder, list_s3_directory};
+use risingwave_connector::source::iceberg::{
+    get_parquet_fields, list_s3_directory, new_s3_operator,
+};
 pub use risingwave_pb::expr::table_function::PbType as TableFunctionType;
 use risingwave_pb::expr::PbTableFunction;
 use thiserror_ext::AsReport;
@@ -181,10 +183,17 @@ impl TableFunction {
 
                 let schema = tokio::task::block_in_place(|| {
                     RUNTIME.block_on(async {
-                        let parquet_stream_builder = create_parquet_stream_builder(
+                        let op = new_s3_operator(
                             eval_args[2].clone(),
                             eval_args[3].clone(),
                             eval_args[4].clone(),
+                            match files.as_ref() {
+                                Some(files) => files[0].clone(),
+                                None => eval_args[5].clone(),
+                            },
+                        )?;
+                        let fields = get_parquet_fields(
+                            op,
                             match files.as_ref() {
                                 Some(files) => files[0].clone(),
                                 None => eval_args[5].clone(),
@@ -193,7 +202,7 @@ impl TableFunction {
                         .await?;
 
                         let mut rw_types = vec![];
-                        for field in parquet_stream_builder.schema().fields() {
+                        for field in &fields {
                             rw_types.push((
                                 field.name().to_string(),
                                 IcebergArrowConvert.type_from_field(field)?,
